Name event parameters in the present tense
The parameters in composable functions that send events are typically
named on + verb in the present tense, like in the very common examples
in Compose foundation code: onClick or onTextChange. It a common
mistake to use the past tense in some of these events, so for
consistency’s sake, we'll want to adjust the tense of the verbs to
present.

Related rule: compose:parameter-naming

https://mrmans0n.github.io/compose-rules/rules/#naming-parameters-properly